Move optional requires into their related methods
This way if the user isn't using the router, or a settings file, we don't waste cycles loading the related classes.

This is particularly important for the router, since hanami-router is not a fixed dependency of the hamami gem (i.e. it's not in the gemspec, and will instead be put into the generated apps' Gemfiles), so if the router gem is not available at all, attempting to load it at the top of lib/hanami/application.rb will result in an unwanted LoadError. This change fixes that.